move function to crypto
